third_party/blink/web_tests/platform/generic/fast/forms/submit-add-remove-element-expected.txt

Test that form.submit() when submitted from JS, form element added and removed, should submit the form without added element

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS event.target.contentWindow.location.search is "?textData=b&submitform=Step"
PASS successfullyParsed is true

TEST COMPLETE

 
